refactor: added usage of sync api

+++ package/src/store/apis/deleteReactions.ts
@@ -1,5 +1,19 @@
 import { QuickSqliteClient } from '../QuickSqliteClient';
+import { createDeleteQuery } from '../sqlite-utils/createDeleteQuery';
 
-export const deleteReactionsForMessage = ({ messageId }: { messageId: string }) => {
-  QuickSqliteClient.executeSql(`DELETE FROM reactions where messageId = ?`, [messageId]);
+export const deleteReactionsForMessage = ({
+  flush = true,
+  messageId,
+}: {
+  messageId: string;
+  flush?: boolean;
+}) => {
+  const query = createDeleteQuery('reactions', {
+    messageId,
+  });
+  if (flush) {
+    QuickSqliteClient.executeSql(`DELETE FROM reactions where messageId = ?`, [messageId]);
+  }
+
+  return [query];
 };
